Delete Device Tests adjustments

+++ backend/api/clientEventHandlers/ClientWantsToDeleteDevice.cs
@@ -1,4 +1,5 @@
 ﻿using System.ComponentModel.DataAnnotations;
+using System.Data.SqlTypes;
 using api.ClientEventFilters;
 using api.helpers;
 using api.serverEventModels;
@@ -38,8 +39,16 @@ public class ClientWantsToDeleteDevice : BaseEventHandler<ClientWantsToDeleteDev
             throw new AccessViolationException("You do not have permission to delete this device");
         }
 
-        _deviceReadingsService.DeleteAllReadings(dto.Id);
-        _deviceService.DeleteDevice(dto.Id);
+        if (!_deviceReadingsService.DeleteAllReadings(dto.Id))
+        {
+            throw new SqlTypeException("Something went wrong when deleting readings for device #" + dto.Id);
+        }
+
+        if (!_deviceService.DeleteDevice(dto.Id))
+        {
+            throw new SqlTypeException("Something went wrong when deleting device #" + dto.Id);
+        }
+        
         //removes the device from stateService
         StateService.RemoveUserFromDevice(dto.Id, socket.ConnectionInfo.Id);
